requirements.txt
discord.py
pyyaml
lxml
requests

